ДЗ 5: Получил погоду с сервера

+++ app/src/main/java/com/example/myappweather/repository/dto/InfoDTO.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 
 data class InfoDTO(
     @SerializedName("lat")
-    val lat: Int,
+    val lat: Double,
     @SerializedName("lon")
-    val lon: Int,
+    val lon: Double,
     @SerializedName("url")
     val url: String
 )
