+++ app/src/main/java/com/example/myappweather/repository/dto/PartDTO.kt
@@ -19,7 +19,7 @@ data class PartDTO(
     @SerializedName("polar")
     val polar: Boolean,
     @SerializedName("prec_mm")
-    val precMm: Int,
+    val precMm: Double,
     @SerializedName("prec_period")
     val precPeriod: Int,
     @SerializedName("prec_prob")
